Wrong size read causing hangup.
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/trunk@366398 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-core/src/main/java/org/apache/activemq/openwire/DataStreamMarshaller.java
@@ -175,7 +175,7 @@ abstract public class DataStreamMarshaller {
                     }
                     o.setStackTrace(ss);
                 } else {
-                    int size = dataIn.readInt();
+                    short size = dataIn.readShort();
                     for (int i = 0; i < size; i++) {
                         readString(dataIn, bs);
                         readString(dataIn, bs);
